files_ms/requirements.txt
flask==2.0.3
Werkzeug==2.0.3
jinja2==3.0.2
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==0.19.2
Flask-Minify==0.37
requests==2.28.2
auth0-python